Client/requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
audioread==3.0.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
Faker==17.5.0
Flask==2.2.3
Flask-SocketIO==5.3.2
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.5.3
msgpack==1.0.5
numba==0.56.4
numpy==1.21.6
packaging==23.0
Pillow==9.4.0
pooch==1.6.0
pycparser==2.21
pymongo==4.3.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.5.1
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.15.0
